refactor: split metrics out
